refactor(test): P5-1 follow-up — extract canvas2d Wave 2 cluster
Continue test_widget_bridge.cpp split. Extracts the 555-line "Wave 2
canvas2d cheap wiring" cluster — pulp #1638 baseline-corruption + the
five compat.json partial → supported entries (fill/clip evenodd,
roundRect 4-corner radii, ellipse rotation, strokeText dedicated
command) — into a sibling TU.

* test_widget_bridge.cpp: 12,443 → 11,888 lines
* New target pulp-test-widget-bridge-canvas2d-wave2: 615 lines,
  10 TEST_CASEs, 74 assertions

No behaviour change; identical TEST_CASE bodies, helpers (canvasFromBridge)
duplicated to keep the split self-contained per the existing
test_widget_bridge_canvas2d.cpp pattern.
