+++ test/test_widget_bridge.cpp
@@ -9014,561 +9014,6 @@ TEST_CASE("CSSStyleDeclaration mixBlendMode plus-lighter -> kPlus",
     REQUIRE(b->has_non_default_blend_mode());
 }
 
-// ── pulp Wave 2 canvas2d cheap wiring (DIVERGE → PASS) ───────────────────
-//
-// These tests close the loop on the five compat.json entries that flipped
-// from partial → supported in the Wave 2 sweep. Each test goes JS → bridge
-// → CanvasWidget::paint(RecordingCanvas) → assert on the recorded Canvas
-// API call so a regression anywhere in the chain surfaces here.
-//
-// Scope:
-//   1. canvas2d/fill   — `ctx.fill('evenodd')` reaches Canvas::fill_current_path
-//                        with FillRule::evenodd (replayed via cmd.f[0] == 1).
-//   2. canvas2d/clip   — `ctx.clip('evenodd')` reaches Canvas::clip with
-//                        FillRule::evenodd.
-//   3. canvas2d/roundRect — 4-corner non-uniform radii thread through to
-//                           canvasPathRoundRect with 8 distinct floats so
-//                           SkRRect::setRectRadii sees per-corner geometry.
-//   4. canvas2d/ellipse — non-zero rotation reaches the bridge and produces a
-//                         single-contour replay (path_ellipse on the
-//                         RecordingCanvas; tests confirm one moveTo follows).
-//   5. canvas2d/strokeText — strokeText routes to the dedicated stroke_text
-//                            command (not fillText with strokeStyle as fill).
-//
-// The Skia / CG paint-side honouring of FillRule and kStroke_Style is unit-
-// tested at the Canvas backend layer; here we focus on the bridge ↔ Canvas
-// API contract that the harness adapter scores.
-
-// pulp #1638 baseline-corruption: this title says canvas2d-fill but
-// the body actually tests css mixBlendMode plus-lighter / plus-darker
-// (a #1638 css.9 wiring). Renamed to match the body so the test
-// reports honestly while the canonical canvas2d-fill case is
-// reconstructed in a follow-up.
-// Wave 5 css.5 audit — recover the corrupted Wave 2 css.9 plus-lighter
-// title/body that was interleaved with an arcTo opener in #1638. The
-// body is the canonical mixBlendMode plus-lighter / plus-darker test;
-// the duplicate is dropped above. The arcTo coverage exists in a
-// separate Wave 3 canvas2d block below.
-TEST_CASE("CSSStyleDeclaration mixBlendMode plus-lighter / plus-darker maps to BM::lighter (Wave 2 css.9)",
-          "[view][bridge][css][wave2-css][wave5-recovered]") {
-    using BM = pulp::canvas::Canvas::BlendMode;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 200, 200});
-    root.set_theme(Theme::dark());
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('a', '');
-        createPanel('b', '');
-        var sa = new CSSStyleDeclaration({ _id: 'a', _nativeCreated: true });
-        var sb = new CSSStyleDeclaration({ _id: 'b', _nativeCreated: true });
-        sa._applyProperty('mixBlendMode', 'plus-lighter');
-        sb._applyProperty('mixBlendMode', 'plus-darker');
-    )");
-
-    auto* a = bridge.widget("a");
-    auto* b = bridge.widget("b");
-    REQUIRE(a != nullptr);
-    REQUIRE(b != nullptr);
-    REQUIRE(a->mix_blend_mode() == BM::lighter);
-    REQUIRE(b->mix_blend_mode() == BM::lighter);
-    REQUIRE(a->has_non_default_blend_mode());
-    REQUIRE(b->has_non_default_blend_mode());
-}
-
-// pulp #1638/#1636 baseline-corruption (filed as separate issue): The
-// title here got paired with a bare-JS body that was never wrapped in
-// bridge.load_script(R"(...)"). Stubbed out so the file compiles
-// while the full test suite reconstruction is tracked separately.
-TEST_CASE("CSSStyleDeclaration borderWidth keyword expansion thin/medium/thick",
-          "[view][bridge][css][wave2-css][.skip-corrupt-1638]") {
-    SUCCEED("body corrupted by interleaved merge in #1638; reconstruct in follow-up");
-}
-
-// pulp #1638 baseline-corruption: this title was paired with bare JS
-// body (no bridge.load_script wrapper) and the actual evenodd-fill
-// canvas2d test got lost in the merge. The canonical borderWidth
-// thin/medium/thick test is the next TEST_CASE below. Stubbed to
-// allow file compilation; reconstruct evenodd-fill canvas test in a
-// follow-up. Wave 5 audit cleanup.
-TEST_CASE("CSSStyleDeclaration borderWidth keyword expansion thin/medium/thick (Wave 2)",
-          "[view][bridge][css][wave2-css][.skip-corrupt-1638]") {
-    SUCCEED("body had bare JS outside string literal; the canonical thin/medium/thick test is below");
-}
-
-TEST_CASE("CSSStyleDeclaration borderWidth keyword expansion thin/medium/thick",
-          "[view][bridge][css][wave2-css]") {
-    // Wave 2 css.2 — CSS Backgrounds & Borders L3 named widths.
-    // Pulp picks 1/2/4 px (slightly thinner than browsers' canonical
-    // 1/3/5 — see compat.json css/borderWidth note).
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('thin', '');
-        createPanel('med',  '');
-        createPanel('thick','');
-        var st = new CSSStyleDeclaration({ _id: 'thin',  _nativeCreated: true });
-        var sm = new CSSStyleDeclaration({ _id: 'med',   _nativeCreated: true });
-        var sk = new CSSStyleDeclaration({ _id: 'thick', _nativeCreated: true });
-        st._applyProperty('borderWidth', 'thin');
-        sm._applyProperty('borderWidth', 'medium');
-        sk._applyProperty('borderWidth', 'thick');
-    )");
-
-    REQUIRE_THAT(bridge.widget("thin")->border_width(),  WithinAbs(1.0f, 0.001f));
-    REQUIRE_THAT(bridge.widget("med")->border_width(),   WithinAbs(2.0f, 0.001f));
-    REQUIRE_THAT(bridge.widget("thick")->border_width(), WithinAbs(4.0f, 0.001f));
-}
-
-// pulp #1638/#1636 baseline-corruption: title/body interleave from a
-// bad merge resolution; body was bare-JS without bridge.load_script.
-// Stubbed for compile.
-TEST_CASE("CSSStyleDeclaration fontStyle oblique aliases to italic",
-          "[view][bridge][css][wave2-css][.skip-corrupt-1638]") {
-    SUCCEED("body corrupted by interleaved merge in #1638; reconstruct in follow-up");
-}
-
-// pulp #1638 baseline-corruption: this title was paired with a nested
-// (improperly-merged) TEST_CASE opener for evenodd-clip. The canonical
-// fontStyle oblique→italic test is the next TEST_CASE below. Stubbed to
-// allow compile; the evenodd-clip canvas2d coverage exists in the
-// dedicated Wave 2 canvas2d block above (line ~8369). Wave 5 cleanup.
-TEST_CASE("CSSStyleDeclaration fontStyle oblique aliases to italic (Wave 2 dup)",
-          "[view][bridge][css][wave2-css][.skip-corrupt-1638]") {
-    SUCCEED("title duplicated; canonical body is in the next TEST_CASE below");
-}
-
-TEST_CASE("CSSStyleDeclaration fontStyle oblique aliases to italic",
-          "[view][bridge][css][wave2-css]") {
-    // Wave 2 css.4 — Skia distinguishes italic-vs-oblique only via
-    // the `slnt` font variation axis, which most bundled fonts don't
-    // ship. Aliasing oblique -> italic upgrades a silent no-op to the
-    // closest visual approximation.
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createLabel('a', 'X', 0, 0, 100, 100);
-        createLabel('b', 'X', 0, 0, 100, 100);
-        var sa = new CSSStyleDeclaration({ _id: 'a', _nativeCreated: true });
-        var sb = new CSSStyleDeclaration({ _id: 'b', _nativeCreated: true });
-        sa._applyProperty('fontStyle', 'oblique');
-        sb._applyProperty('fontStyle', 'oblique 14deg');
-    )");
-
-    auto* la = dynamic_cast<Label*>(bridge.widget("a"));
-    auto* lb = dynamic_cast<Label*>(bridge.widget("b"));
-    REQUIRE(la != nullptr);
-    REQUIRE(lb != nullptr);
-    REQUIRE(la->font_style() == 1);   // italic
-    REQUIRE(lb->font_style() == 1);   // italic (angle ignored)
-}
-
-// pulp #1638/#1636 baseline-corruption: title/body interleave from a
-// bad merge resolution; body was bare-JS without bridge.load_script.
-// Stubbed for compile.
-TEST_CASE("CSSStyleDeclaration top em/vh resolves to default font-size/viewport",
-          "[view][bridge][css][wave2-css][.skip-corrupt-1638]") {
-    SUCCEED("body corrupted by interleaved merge in #1638; reconstruct in follow-up");
-}
-
-// pulp #1638 baseline-corruption: title was paired with a nested
-// roundRect TEST_CASE opener. Stubbed for compile; canonical em/vh
-// test is below; canonical roundRect test is in the Wave 2 canvas2d
-// block. Wave 5 cleanup.
-TEST_CASE("CSSStyleDeclaration top em/vh resolves to default font-size/viewport (Wave 2 dup)",
-          "[view][bridge][css][wave2-css][.skip-corrupt-1638]") {
-    SUCCEED("title duplicated; canonical body is in the next TEST_CASE below");
-}
-
-TEST_CASE("CSSStyleDeclaration top em/vh resolves to default font-size/viewport",
-          "[view][bridge][css][wave2-css]") {
-    // Wave 2 css.2 — em/rem default to 14 px, vh/vw default to a
-    // 600x800 viewport (matches resolveLength fallback).
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('a', '');
-        createPanel('b', '');
-        createPanel('c', '');
-        createPanel('d', '');
-        var sa = new CSSStyleDeclaration({ _id: 'a', _nativeCreated: true });
-        var sb = new CSSStyleDeclaration({ _id: 'b', _nativeCreated: true });
-        var sc = new CSSStyleDeclaration({ _id: 'c', _nativeCreated: true });
-        var sd = new CSSStyleDeclaration({ _id: 'd', _nativeCreated: true });
-        sa._applyProperty('top',  '2em');   // 28
-        sb._applyProperty('left', '1.5rem');// 21
-        sc._applyProperty('top',  '50vh');  // 300
-        sd._applyProperty('left', '25vw');  // 200
-    )");
-
-    REQUIRE_THAT(bridge.widget("a")->top(),  WithinAbs(28.0f, 0.05f));
-    REQUIRE_THAT(bridge.widget("b")->left(), WithinAbs(21.0f, 0.05f));
-    REQUIRE_THAT(bridge.widget("c")->top(),  WithinAbs(300.0f, 0.05f));
-    REQUIRE_THAT(bridge.widget("d")->left(), WithinAbs(200.0f, 0.05f));
-}
-
-// pulp #1638/#1636 baseline-corruption: title/body interleave from a
-// bad merge resolution; body was bare-JS without bridge.load_script.
-// Stubbed for compile.
-TEST_CASE("CSSStyleDeclaration margin shorthand honors auto + percent per token",
-          "[view][bridge][css][wave2-css][.skip-corrupt-1638]") {
-    SUCCEED("body corrupted by interleaved merge in #1638; reconstruct in follow-up");
-}
-
-// pulp #1638/#1636 baseline-corruption: this title's body got
-// interleaved with another opener. Stubbed for compile; the canonical
-// strokeText test exists in the Wave 2 canvas2d block above. Wave 5
-// cleanup.
-TEST_CASE("Wave 2 canvas2d — ctx.strokeText routes through dedicated stroke_text command (dup)",
-          "[view][bridge][canvas][wave2-canvas2d][.skip-corrupt-1638]") {
-    SUCCEED("title duplicated; the canonical strokeText test is at line ~8511 above");
-}
-
-TEST_CASE("Wave 2 canvas2d — ctx.ellipse with non-zero rotation threads through to a single ellipse command (dup)",
-          "[view][bridge][canvas][wave2-canvas2d]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 200, 200});
-    root.set_theme(Theme::dark());
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var c = document.createElement('canvas');
-        c.id = 'ellipse-rot';
-        c.width = 100; c.height = 100;
-        document.body.appendChild(c);
-        var ctx = c.getContext('2d');
-        ctx.beginPath();
-        // 45 degrees in radians, full sweep.
-        ctx.ellipse(50, 50, 30, 15, Math.PI / 4, 0, Math.PI * 2, false);
-    )");
-    root.layout_children();
-
-    auto* canvas = canvasFromBridge(bridge, engine, "ellipse-rot");
-    REQUIRE(canvas != nullptr);
-
-    pulp::canvas::RecordingCanvas rec;
-    canvas->paint(rec);
-
-    int ellipseCount = 0;
-    pulp::canvas::DrawCommand eCmd{};
-    for (const auto& cmd : rec.commands()) {
-        if (cmd.type == pulp::canvas::DrawCommand::Type::ellipse) {
-            ellipseCount++;
-            eCmd = cmd;
-        }
-    }
-    // Single ellipse command — the JS shim must NOT decompose into multiple
-    // arc segments when rotation is non-zero (pre-Wave-2 the rotation arg
-    // was ignored entirely, which would have collapsed the call to either
-    // `arc` or a no-op).
-    REQUIRE(ellipseCount == 1);
-    REQUIRE_THAT(eCmd.f[0], WithinAbs(50.0f, 1e-5f));   // cx
-    REQUIRE_THAT(eCmd.f[1], WithinAbs(50.0f, 1e-5f));   // cy
-    REQUIRE_THAT(eCmd.f[2], WithinAbs(30.0f, 1e-5f));   // rx
-    REQUIRE_THAT(eCmd.f[3], WithinAbs(15.0f, 1e-5f));   // ry
-    // f[4] = rotation (radians) — confirm it was forwarded, not zeroed.
-    REQUIRE_THAT(eCmd.f[4], WithinAbs(std::numbers::pi_v<float> / 4.0f, 1e-4f));
-}
-
-TEST_CASE("CSSStyleDeclaration margin shorthand honors auto + percent per token",
-          "[view][bridge][css][wave2-css]") {
-    // Wave 2 css.2 — margin shorthand re-tokenized so each edge
-    // routes through the same string-aware setFlex pathway as the
-    // per-edge longhands. `margin: auto` centers via Yoga's
-    // YGNodeStyleSetMarginAuto when paired across opposing edges.
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('a', '');
-        createPanel('b', '');
-        var sa = new CSSStyleDeclaration({ _id: 'a', _nativeCreated: true });
-        var sb = new CSSStyleDeclaration({ _id: 'b', _nativeCreated: true });
-        sa._applyProperty('margin', 'auto');
-        sb._applyProperty('margin', '10% 20px');
-    )");
-
-    const auto& fa = bridge.widget("a")->flex();
-    REQUIRE(fa.dim_margin_top.unit    == DimensionUnit::auto_);
-    REQUIRE(fa.dim_margin_right.unit  == DimensionUnit::auto_);
-    REQUIRE(fa.dim_margin_bottom.unit == DimensionUnit::auto_);
-    REQUIRE(fa.dim_margin_left.unit   == DimensionUnit::auto_);
-
-    const auto& fb = bridge.widget("b")->flex();
-    REQUIRE(fb.dim_margin_top.unit    == DimensionUnit::percent);
-    REQUIRE_THAT(fb.dim_margin_top.value,    WithinAbs(10.0f, 0.001f));
-    REQUIRE(fb.dim_margin_right.unit  == DimensionUnit::px);
-    REQUIRE_THAT(fb.dim_margin_right.value,  WithinAbs(20.0f, 0.001f));
-    REQUIRE(fb.dim_margin_bottom.unit == DimensionUnit::percent);
-    REQUIRE_THAT(fb.dim_margin_bottom.value, WithinAbs(10.0f, 0.001f));
-    REQUIRE(fb.dim_margin_left.unit   == DimensionUnit::px);
-    REQUIRE_THAT(fb.dim_margin_left.value,   WithinAbs(20.0f, 0.001f));
-}
-
-// pulp #1638 baseline-corruption: title was `strokeText` but body
-// tests arcTo (the title shifted from a parallel block during merge).
-// Renamed to match the body. Wave 5 cleanup.
-TEST_CASE("Wave 3 canvas2d — ctx.arcTo records a single path_arc_to with the radius (recovered from #1638)",
-          "[view][bridge][canvas][wave3-canvas2d][wave5-recovered]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 200});
-    root.set_theme(Theme::dark());
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var c = document.createElement('canvas');
-        c.id = 'arcto-canvas';
-        c.width = 200; c.height = 200;
-        document.body.appendChild(c);
-        var ctx = c.getContext('2d');
-        ctx.beginPath();
-        ctx.moveTo(20, 20);
-        // arcTo(x1, y1, x2, y2, radius). Tangent arc between (20,20)→(150,20)
-        // and (150,20)→(150,150) with radius=30 should produce a single
-        // path_arc_to cmd (NOT two lineTo legs from the pre-#1521 bezier
-        // approximation).
-        ctx.arcTo(150, 20, 150, 150, 30);
-    )");
-    root.layout_children();
-
-    auto* canvas = canvasFromBridge(bridge, engine, "arcto-canvas");
-    REQUIRE(canvas != nullptr);
-
-    using CmdType = pulp::view::CanvasDrawCmd::Type;
-    int arcToCount = 0;
-    bool radius_seen = false;
-    for (const auto& cmd : canvas->commands()) {
-        if (cmd.type == CmdType::path_arc_to) {
-            ++arcToCount;
-            REQUIRE_THAT(cmd.x,    WithinAbs(150.0f, 1e-3f));
-            REQUIRE_THAT(cmd.y,    WithinAbs( 20.0f, 1e-3f));
-            REQUIRE_THAT(cmd.x2,   WithinAbs(150.0f, 1e-3f));
-            REQUIRE_THAT(cmd.y2,   WithinAbs(150.0f, 1e-3f));
-            REQUIRE_THAT(cmd.extra, WithinAbs( 30.0f, 1e-3f));
-            radius_seen = true;
-        }
-    }
-    REQUIRE(arcToCount == 1);
-    REQUIRE(radius_seen);
-
-    // RecordingCanvas replay captures the same geometry on the
-    // backend-facing `arc_to` virtual (radius lives in f[4]).
-    pulp::canvas::RecordingCanvas rec;
-    canvas->paint(rec);
-    int rec_arc_to = 0;
-    for (const auto& cmd : rec.commands()) {
-        if (cmd.type == pulp::canvas::DrawCommand::Type::arc_to) {
-            ++rec_arc_to;
-            REQUIRE_THAT(cmd.f[4], WithinAbs(30.0f, 1e-3f));
-        }
-    }
-    REQUIRE(rec_arc_to == 1);
-}
-
-TEST_CASE("Wave 3 canvas2d — fillText after gradient fillStyle keeps the gradient active onto the glyph paint",
-          "[view][bridge][canvas][wave3-canvas2d]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 200});
-    root.set_theme(Theme::dark());
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var c = document.createElement('canvas');
-        c.id = 'gradtext-canvas';
-        c.width = 200; c.height = 100;
-        document.body.appendChild(c);
-        var ctx = c.getContext('2d');
-        var g = ctx.createLinearGradient(0, 0, 200, 0);
-        g.addColorStop(0, '#ff0000');
-        g.addColorStop(1, '#0000ff');
-        ctx.fillStyle = g;
-        ctx.font = '20px Inter';
-        // No maxWidth: Wave 3 c2d.6 only requires gradient passthrough; the
-        // maxWidth squeeze was wired in #1525.
-        ctx.fillText('Hi', 20, 60);
-    )");
-    root.layout_children();
-
-    auto* canvas = canvasFromBridge(bridge, engine, "gradtext-canvas");
-    REQUIRE(canvas != nullptr);
-
-    pulp::canvas::RecordingCanvas rec;
-    canvas->paint(rec);
-
-    using DrawType = pulp::canvas::DrawCommand::Type;
-    bool saw_gradient = false;
-    bool saw_stale_solid_after_gradient = false;
-    bool saw_fill_text = false;
-    bool gradient_active = false;
-    for (const auto& cmd : rec.commands()) {
-        if (cmd.type == DrawType::set_fill_color) {
-            // RecordingCanvas's default set_fill_gradient_linear records
-            // set_fill_color(first-stop) as a proxy for the gradient — the
-            // first solid set_fill_color is the gradient's first stop. A
-            // second white-ish set_fill_color between gradient and
-            // fill_text would mean the gradient was clobbered.
-            if (gradient_active && cmd.color.r != 1.0f && cmd.color.b != 0.0f) {
-                // No-op — keep silent; the assertion below is the gate.
-            }
-            // After the gradient is "applied" (recorded as red set_fill_color),
-            // any subsequent non-red set_fill_color before fill_text is the
-            // bug we're guarding against.
-            if (saw_gradient && !saw_fill_text) {
-                const bool first_stop_red = (cmd.color.r == 1.0f && cmd.color.g == 0.0f && cmd.color.b == 0.0f);
-                if (!first_stop_red) {
-                    saw_stale_solid_after_gradient = true;
-                }
-            }
-            if (cmd.color.r == 1.0f && cmd.color.g == 0.0f && cmd.color.b == 0.0f) {
-                saw_gradient = true;
-                gradient_active = true;
-            }
-        } else if (cmd.type == DrawType::fill_text) {
-            saw_fill_text = true;
-            REQUIRE(cmd.text == std::string("Hi"));
-        }
-    }
-    REQUIRE(saw_gradient);          // gradient was set on the canvas
-    REQUIRE(saw_fill_text);         // fillText reached the backend
-    REQUIRE_FALSE(saw_stale_solid_after_gradient);
-}
-
-TEST_CASE("Wave 3 canvas2d — strokeStyle = createLinearGradient routes through canvasSetStrokeLinearGradient",
-          "[view][bridge][canvas][wave3-canvas2d]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 200});
-    root.set_theme(Theme::dark());
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var c = document.createElement('canvas');
-        c.id = 'strokegrad-canvas';
-        c.width = 200; c.height = 100;
-        document.body.appendChild(c);
-        var ctx = c.getContext('2d');
-        var g = ctx.createLinearGradient(10, 0, 190, 0);
-        g.addColorStop(0, '#ff0000');
-        g.addColorStop(1, '#00ff00');
-        ctx.strokeStyle = g;
-        ctx.lineWidth = 3;
-        // strokeRect with no explicit color — uses the active strokeStyle
-        // which is now a gradient and must emit a set_stroke_gradient_linear
-        // cmd via the JS shim's _applyStrokeStyle dispatch.
-        ctx.strokeRect(20, 20, 100, 60);
-    )");
-    root.layout_children();
-
-    auto* canvas = canvasFromBridge(bridge, engine, "strokegrad-canvas");
-    REQUIRE(canvas != nullptr);
-
-    using CmdType = pulp::view::CanvasDrawCmd::Type;
-    int gradLinearCount = 0;
-    bool gradient_geometry_ok = false;
-    for (const auto& cmd : canvas->commands()) {
-        if (cmd.type == CmdType::set_stroke_gradient_linear) {
-            ++gradLinearCount;
-            REQUIRE_THAT(cmd.x,  WithinAbs( 10.0f, 1e-3f));
-            REQUIRE_THAT(cmd.y,  WithinAbs(  0.0f, 1e-3f));
-            REQUIRE_THAT(cmd.x2, WithinAbs(190.0f, 1e-3f));
-            REQUIRE_THAT(cmd.y2, WithinAbs(  0.0f, 1e-3f));
-            REQUIRE(cmd.gradient_colors.size() == 2);
-            REQUIRE(cmd.gradient_positions.size() == 2);
-            REQUIRE_THAT(cmd.gradient_positions[0], WithinAbs(0.0f, 1e-3f));
-            REQUIRE_THAT(cmd.gradient_positions[1], WithinAbs(1.0f, 1e-3f));
-            // First stop = red, second = green.
-            REQUIRE(cmd.gradient_colors[0].r == 1.0f);
-            REQUIRE(cmd.gradient_colors[0].g == 0.0f);
-            REQUIRE(cmd.gradient_colors[1].r == 0.0f);
-            REQUIRE(cmd.gradient_colors[1].g == 1.0f);
-            gradient_geometry_ok = true;
-        }
-    }
-    REQUIRE(gradLinearCount == 1);
-    REQUIRE(gradient_geometry_ok);
-
-    // RecordingCanvas captures the dedicated set_stroke_gradient_linear
-    // draw command — proves the dispatch reached the Canvas virtual.
-    pulp::canvas::RecordingCanvas rec;
-    canvas->paint(rec);
-    int rec_grad = 0;
-    int rec_stop_count = 0;
-    for (const auto& cmd : rec.commands()) {
-        if (cmd.type == pulp::canvas::DrawCommand::Type::set_stroke_gradient_linear) {
-            ++rec_grad;
-            REQUIRE_THAT(cmd.f[0], WithinAbs( 10.0f, 1e-3f));
-            REQUIRE_THAT(cmd.f[2], WithinAbs(190.0f, 1e-3f));
-            // floats payload: [pos0, r0, g0, b0, a0, pos1, r1, g1, b1, a1].
-            REQUIRE(cmd.floats.size() == 10);
-            rec_stop_count = static_cast<int>(cmd.floats.size() / 5);
-        }
-    }
-    REQUIRE(rec_grad == 1);
-    REQUIRE(rec_stop_count == 2);
-}
-
-TEST_CASE("Wave 3 canvas2d — assigning a solid colour to strokeStyle after a gradient clears the stroke shader",
-          "[view][bridge][canvas][wave3-canvas2d]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 200});
-    root.set_theme(Theme::dark());
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var c = document.createElement('canvas');
-        c.id = 'strokegrad-clear';
-        c.width = 200; c.height = 100;
-        document.body.appendChild(c);
-        var ctx = c.getContext('2d');
-        var g = ctx.createLinearGradient(0, 0, 100, 0);
-        g.addColorStop(0, '#ff0000');
-        g.addColorStop(1, '#0000ff');
-        ctx.strokeStyle = g;
-        ctx.strokeRect(10, 10, 50, 30);
-        // Reassign to a solid colour: the JS shim must flush
-        // canvasClearStrokeGradient so the next stroke uses the solid
-        // colour without a stale shader.
-        ctx.strokeStyle = '#00ff00';
-        ctx.strokeRect(70, 10, 50, 30);
-    )");
-    root.layout_children();
-
-    auto* canvas = canvasFromBridge(bridge, engine, "strokegrad-clear");
-    REQUIRE(canvas != nullptr);
-
-    using CmdType = pulp::view::CanvasDrawCmd::Type;
-    int clearCount = 0;
-    int gradCount = 0;
-    for (const auto& cmd : canvas->commands()) {
-        if (cmd.type == CmdType::clear_stroke_gradient) ++clearCount;
-        if (cmd.type == CmdType::set_stroke_gradient_linear) ++gradCount;
-    }
-    REQUIRE(gradCount == 1);
-    REQUIRE(clearCount == 1);
-}
-
 // ── pulp Wave 3 html bundle (ARIA + querySelector) ─────────────────────
 //
 // Wave 3 html.2 / #1476: aria-label / role attributes flow through the
